+++ src/modules/nodes/NodesTypes.ts
@@ -5,6 +5,7 @@ export interface NodeType {
 
 export interface NodeInfoType {
     host: string;
+    version: string;
     channel: string;
     config: {
         model_export_enabled: boolean;