+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Add CSV export (#117)
 -   Compute Plans comparison page (#130)
 -   Owner section in compute plan sider (#140)
+-   Display backend version (#143)
 
 ### Fixed